refactor to converters package
